publisher notification uses base
