+++ app/controllers/concerns/publishers/login_required.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Publishers
+  module LoginRequired
+    extend ActiveSupport::Concern
+
+    included do
+      include Authenticated
+
+      self.authentication_scope = :publisher
+    end
+  end
+end
