+++ app/controllers/publishers/notifications_controller.rb
@@ -1,23 +1,9 @@
-class Publishers::NotificationsController < Publishers::BaseController
-  NOTIFICATIONS_PER_PAGE = 30
+module Publishers
+  class NotificationsController < ::NotificationsController
+    include LoginRequired
 
-  before_action :load_notifications
-  after_action :mark_notifications_as_read
-
-  def index
-    @unread_count = @raw_notifications.unread.count
-    @pagy, @notifications = pagy(@raw_notifications, items: NOTIFICATIONS_PER_PAGE)
-  end
-
-  private
-
-  def load_notifications
-    @raw_notifications = current_publisher.notifications
-                                        .where(created_at: (Time.current - DATA_ACCESS_PERIOD_FOR_PUBLISHERS)..)
-                                        .newest_first
-  end
-
-  def mark_notifications_as_read
-    @notifications.mark_as_read
+    def notification_user
+      current_publisher
+    end
   end
 end
