+++ config/locales/notifications.yml
@@ -47,6 +47,14 @@ en:
       reference_received: reference for a candidate
     jobseekers/self_disclosure_request_received_notification:
       message_html: You have received a request to complete a %{link}
+    index:
+      heading: Notifications
+      no_notifications: You have no notifications.
+      skip_link: Skip to notifications
+      title:
+        zero: No unread notifications
+        one: 1 unread notification
+        other: "%{count} unread notifications"
 
   errors:
     jobs:
@@ -64,17 +72,6 @@ en:
       already_exists: You're already subscribed to an alert for these search criteria
 
   jobseekers:
-    notifications:
-      index:
-        heading: Notifications
-        no_notifications: You have no notifications.
-        skip_link: Skip to notifications
-        title:
-          zero: No unread notifications
-          one: 1 unread notification
-          other: "%{count} unread notifications"
-
-  publishers:
     notifications:
       index:
         heading: Notifications
